requirements.txt
alembic==1.4.3
aniso8601==8.0.0
blinker==1.4
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.1.0
Mako==1.1.3
MarkupSafe==1.1.1
polyline==1.4.0
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
requests==2.25.0
Shapely==1.7.1
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
validate-email==1.3
Werkzeug==1.0.1
